Remove legacy models.py and refactor imports to direct module paths
- Refactor all imports from llm_orc.models to direct module imports
- Update core modules: model_factory, orchestration, ensemble_execution
- Update all test files to use direct imports from models.anthropic, models.base, etc
- Remove legacy models.py compatibility layer (was 0% coverage)
- Fix import issues and maintain proper indentation
- All 228 tests continue to pass
- Cleaner import structure, no unnecessary re-export layer
