+++ examples/shakespeare_einstein_conversation.py
@@ -19,7 +19,7 @@ from pathlib import Path
 # Add src to path for imports
 sys.path.insert(0, str(Path(__file__).parent.parent / "src"))
 
-from llm_orc.models import OllamaModel
+from llm_orc.models.ollama import OllamaModel
 from llm_orc.orchestration import Agent, ConversationOrchestrator
 from llm_orc.roles import RoleDefinition
