+++ src/llm_orc/core/execution/orchestration.py
@@ -5,7 +5,7 @@ from typing import Any
 
 from llm_orc.core.communication.protocol import ConversationManager, MessageProtocol
 from llm_orc.core.config.roles import RoleDefinition
-from llm_orc.models import ModelInterface
+from llm_orc.models.base import ModelInterface
 
 
 class Agent:
