+++ tests/core/auth/test_oauth_token_storage.py
@@ -8,7 +8,7 @@ import pytest
 from llm_orc.core.auth.authentication import CredentialStorage
 from llm_orc.core.config.config_manager import ConfigurationManager
 from llm_orc.core.execution.ensemble_execution import EnsembleExecutor
-from llm_orc.models import OAuthClaudeModel
+from llm_orc.models.anthropic import OAuthClaudeModel
 
 
 class TestOAuthTokenStorage:
