+++ tests/core/execution/test_agent_orchestration.py
@@ -5,7 +5,7 @@ from unittest.mock import AsyncMock, Mock, patch
 import pytest
 
 from llm_orc.core.config.roles import RoleDefinition
-from llm_orc.models import ModelInterface
+from llm_orc.models.base import ModelInterface
 
 
 class TestAgent:
